Added git ignore Rename files
