+++ .gitignore
@@ -1,0 +1,5 @@
+/Debug
+/settings
+/terminal-server.dep
+/terminal-server.ewd
+/terminal-server.ewt
